Приличный код с автотестами

+++ src/main/java/Main.java
@@ -3,25 +3,25 @@ import ru.netology.stats.StatsService;
 public class Main {
     public static void main(String[] args) {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int sumOfAllSales = (int) service.calculateSumOfAllSales(sales);
-        System.out.println("sumOfAllSale:"+sumOfAllSales);
+        long sumOfAllSales = service.calculateSumOfAllSales(sales);
+        System.out.println("sumOfAllSale:" + sumOfAllSales);
 
-        int averageSalePerMonth = (int) service.calculateAverageSalePerMonth (sales);
-        System.out.println("averageSalePerMonth:"+averageSalePerMonth);
+        long averageSalePerMonth = service.calculateAverageSalePerMonth(sales);
+        System.out.println("averageSalePerMonth:" + averageSalePerMonth);
 
-        int numerOfMonthWithMaxSale = (int) service.calculateNumerOfMonthWithMaxSale (sales);
-        System.out.println("numerOfMonthWithMaxSale:"+numerOfMonthWithMaxSale);
+        long numerOfMonthWithMaxSale = service.calculateNumerOfMonthWithMaxSale(sales);
+        System.out.println("numerOfMonthWithMaxSale:" + numerOfMonthWithMaxSale);
 
-        int numerOfMonthWithMinSale = (int) service.calculatenumerOfMonthWithMinSale (sales);
-        System.out.println("numerOfMonthWithMinSale:"+numerOfMonthWithMinSale);
+        long numerOfMonthWithMinSale = service.calculatenumerOfMonthWithMinSale(sales);
+        System.out.println("numerOfMonthWithMinSale:" + numerOfMonthWithMinSale);
 
-        int numberOfMonthesWithUnderAverageSale = (int) service.calculateNumberOfMonthesWithUnderAverageSale (sales);
-        System.out.println("numberOfMonthesWithUnderAverageSale:"+numberOfMonthesWithUnderAverageSale);
+        long numberOfMonthesWithUnderAverageSale = service.calculateNumberOfMonthesWithUnderAverageSale(sales);
+        System.out.println("numberOfMonthesWithUnderAverageSale:" + numberOfMonthesWithUnderAverageSale);
 
-        int numberOfMonthesWithOverAverageSale = (int) service.calculateNumberOfMonthesWithOverAverageSale (sales);
-        System.out.println("numberOfMonthesWithOverAverageSale:"+numberOfMonthesWithOverAverageSale);
+        long numberOfMonthesWithOverAverageSale = service.calculateNumberOfMonthesWithOverAverageSale(sales);
+        System.out.println("numberOfMonthesWithOverAverageSale:" + numberOfMonthesWithOverAverageSale);
     }
